+++ CHANGELOG.md
@@ -1,6 +1,10 @@
 # Change Log
 All notable changes to this project will be documented in this file.
 
+## 2021-02-21
+### Added
+- added `self.default_atom_values` to atoms
+
 ## 2021-02-04
 ### Added
 - added `Rails.application.config.folio_console_ability_lambda` for easier console ability tweaks